pinctrl: renesas: Support RGID
This patch supports adding the RGID.

+++ drivers/pinctrl/renesas/pfc-r8a77951.c
@@ -6055,7 +6055,7 @@ static const struct pinmux_ioctrl_reg pinmux_ioctrl_regs[] = {
 };
 
 static int r8a77951_pin_to_pocctrl(struct sh_pfc *pfc,
-				   unsigned int pin, u32 *pocctrl)
+				   unsigned int pin, u64 *pocctrl)
 {
 	int bit = -EINVAL;
