+++ drivers/pinctrl/renesas/pfc-r8a77965.c
@@ -6262,7 +6262,7 @@ static const struct pinmux_ioctrl_reg pinmux_ioctrl_regs[] = {
 	{ /* sentinel */ },
 };
 
-static int r8a77965_pin_to_pocctrl(struct sh_pfc *pfc, unsigned int pin, u32 *pocctrl)
+static int r8a77965_pin_to_pocctrl(struct sh_pfc *pfc, unsigned int pin, u64 *pocctrl)
 {
 	int bit = -EINVAL;
